requirements.txt
random
Pillow